// src/app/core/models/property-type.enum.ts
export enum PropertyType { 
  APARTMENT = 'APARTMENT',
  HOUSE = 'HOUSE',
  VILLA = 'VILLA',
  COTTAGE = 'COTTAGE',
  CHALET = 'CHALET',
  BUNGALOW = 'BUNGALOW',
  CABIN = 'CABIN',
  STUDIO = 'STUDIO',
  CONDO = 'CONDO',
  PENTHOUSE = 'PENTHOUSE',
  LOFT = 'LOFT',
  FARMHOUSE = 'FARMHOUSE',
  TREEHOUSE = 'TREEHOUSE',
  BOAT = 'BOAT',
  TENT = 'TENT',
  YURT = 'YURT',
  LUXURY_VILLA = 'LUXURY_VILLA',
  BED_AND_BREAKFAST = 'BED_AND_BREAKFAST',
  GUEST_HOUSE = 'GUEST_HOUSE'
}